Create README.md
Terraform: AWS Three-Tier Architecture Design
In this project, we will create a three-tier architecture leveraging Terraform modules to make the process easily repeatable and reusable. Our architecture will reside in a custom VPC.

The web tier will have a bastion host and NAT gateway in the public subnets. The bastion host will serve as our access point to the underlying infrastructure. The NAT Gateway will allow our private subnets to access updates from the internet.

In the application tier, we will create an internet facing load balancer to direct internet traffic to an autoscaling group in the private subnets, along with a backend autoscaling group for our backend application. We will create a script to install the apache webserver in the frontend, and a script to install Node.js in the backend.

We will have another layer of private subnets in the database tier hosting a MySQL database which will eventually access using Node.js. Keep in mind, this is simply an example of infrastructure one could use for a web application.

I will be using Visual Studio Code as my IDE,